requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.5.2
async-timeout==4.0.2
autopep8==2.0.0
billiard==3.6.4.0
cached-property==1.5.2
celery==4.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.16
djangorestframework==3.14.0
flake8==5.0.4
idna==3.4
importlib-metadata==4.2.0
instasent==0.1.4
Jinja2==3.1.2
kombu==4.6.11
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
prompt-toolkit==3.0.32
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
rabbitmq==0.2.0
redmail==0.4.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
vine==1.3.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.10.0
